+++ src/BlockBuilder/FieldType/FieldTypeRegistry.php
@@ -19,6 +19,7 @@ use BlockBuilder\FieldType\Type\LinkFromSitemap\LinkFromSitemapFieldType;
 use BlockBuilder\FieldType\Type\MultipleChoice\MultipleChoiceFieldType;
 use BlockBuilder\FieldType\Type\Number\NumberFieldType;
 use BlockBuilder\FieldType\Type\SingleChoice\SingleChoiceFieldType;
+use BlockBuilder\FieldType\Type\SvgIconPicker\SvgIconPickerFieldType;
 use BlockBuilder\FieldType\Type\Text\TextFieldType;
 use BlockBuilder\FieldType\Type\Textarea\TextareaFieldType;
 use BlockBuilder\FieldType\Type\UserSelector\UserSelectorFieldType;
@@ -62,6 +63,7 @@ final class FieldTypeRegistry
         $this->register(new DatePickerFieldType());
         $this->register(new ColorPickerFieldType());
         $this->register(new IconPickerFieldType());
+        $this->register(new SvgIconPickerFieldType());
         $this->register(new UserSelectorFieldType());
 
         if (count($this->fieldTypesByEnumName) !== count(FieldTypeEnum::cases())) {